real-time-sentiment/requirements.txt
# Core Dependencies
kafka-python==2.0.2
pyspark==3.5.0
streamlit==1.28.1
plotly==5.17.0
pandas==2.1.3
numpy==1.24.3

# Natural Language Processing
textblob==0.17.1
nltk==3.8.1
scikit-learn==1.3.2

# Data Processing
requests==2.31.0
python-json-logger==2.0.7

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Database and Storage
redis==5.0.1
elasticsearch==8.11.0

# Development and Testing
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0

# Optional: Deep Learning
# tensorflow==2.15.0
# torch==2.1.1
# transformers==4.35.2

# Optional: Additional Data Sources
# tweepy==4.14.0
# praw==7.7.1
# newsapi-python==0.2.7
